// src/app/srvapi/models/space-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface SpaceDTO
 */
export interface SpaceDTO {
    /**
     * 
     * @type {number}
     * @memberof SpaceDTO
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof SpaceDTO
     */
    name?: any;
    /**
     * 
     * @type {number}
     * @memberof SpaceDTO
     */
    parentId?: any;
    /**
     * 
     * @type {string}
     * @memberof SpaceDTO
     */
    slug?: any;
}